Rename NetBSD to pkgsrc as suggested by multiple people
The idea behind naming in NetBSD was that it parses NetBSD package list
instead of actualy pkgsrc. My idea was that it may contain less packages
than pkgsrc contains, so I've named it NetBSD to be precise. Well, lets
rename it to pkgsrc as it's likely pretty close to it, and later we'll
switch to actual pkgsrc parsing (it's easy, but INDEX generation takes
quite some time)