Add homebrew parser
It works on `brew info --json=v1 --all' json output and is capable of
parsing package names, versions, homepages and descriptions. Though
homebrew formulae contain download urls, these are unfortunately not
available in the dump.

Currently this dump is not available online, and generating it
ourselves is problematic, so repository is not enabled yet.