Don't use a pipeline in WgetTar fetcher
It's unsafe to use a pipeline since when there's no portable way to
enable an equivalent of bash pipefail. Without this, a command failure
may be missed and treated as success. Here, wget may fail, and this
will silently produce empty output.

So split a pipeline into two commands which use a temporary file.