Refactor fetcher/parser spawning
Implement module factory which is capable of enumerating all
modules of the current package, finding classes by a given
pattern and providing a way to spawn an instance of any saved
class. This simplifies on-demand spawning of *Fetcher and *Parser
classes in RepositoryManager, and remove the need for importing
all modules manually.