Improve branch selection logic
Currently we assign a set of equal versions to a devel branch if and
only if all of them have devel flag. Change this logic to require just
any of them to gave devel flag. This improves handling of the
following situation:

* 2.0rc1 (devel)
* 2.0r1  (non devel, but same version)
* 1.1

Before, 2.0r1 would prevent us from assigning 2.0* to devel branch ->
2.0* would be newest and 1.1 would be outdated, which is not really
correct. Now, 2.0* are devel because just 2.0rc1 is devel, and 1.1 will
remain newest.

This does not affect a lot of metapackages (~50), but most of the
changes are positive. There are a couple of cases like this:

* 2.0b1
* 2.0beta1
* 2.0a1
* 1.0

Where the lack of devel neighbour to 2.0a1 makes it newest erroneously,
but it still may be fixed with .*[ab].* like devel rules.