Add successor flag
This is intended to mark forks or patchsets with supersede original
upstream and are used as a drop-in replacements in some repos. This
flag is the same as `devel` currently, and it works nice (favors both
genuine versions and a newer forks), but in future it may conflict
with original `devel` meaning, which will require implementing more
advanced branches logic. Also, we may want to mark forks specially
in the UI in the future.

But for now, we just use it as an additional metainfo in the rules.

Fixes #496