Introduce metapackages table (#504)
For now this is a replacement for metapackage_repocounts materialized
view, which allows easier management (e.g. adding columns) and will
allow partial updates in the future.

Add new features to it right away: first_seen and last_seen fields
which allow to track metapackage lifetimes. Also add a separate
table for metapackages for which no more packages exist and handle
moving metapackages from and to it.

This table is not used anywhere yet.