Use the `next_page` metadata from crates.io
crates.io will be deprecating its numbered pagination in the near
future. In order to help clients migrate to the new pagination API, a
`next_page` and `prev_page` field is included in the response metadata,
unless you're on the first/last page.

Currently this should not result in any change in functionality (the
values in `next_page` will equal what was being hit before), but once
the new pagination system is in place, this field will switch to using
it, meaning Repology will be unaffected when the old pagination is
eventually removed.

I was not able to test this change locally, as there doesn't seem to be
a way to just run a single fetcher without also having a database set
up, installing C libraries that aren't published to any package manager
that I could find, etc.

Thanks for setting a useful user-agent header so I can reach out with
this PR instead of just breaking your app