Add Gentoo overlay GURU
As stated in [1] "the goal of the GURU project is to create an official
repository of new Gentoo packages that are maintained collaboratively
by Gentoo users".

"GURU is an official Gentoo project, and is therefore bound
by the official Gentoo policies".

The official detailed article about project is available [2] too.

[1] https://wiki.gentoo.org/wiki/Project:GURU
[2] https://dev.gentoo.org/~mgorny/articles/guru-a-new-model-of-contributing-to-gentoo.html