Only allow partial updates
Non-partial update mode was enabled if more that 25% of projects were
modified and was meant as an optimization which reprocesses ALL packages
instead of SIGNIFICANT PART of them, avoiding overhead needed to select
packages to process.

Drop this mode for it doesn't bring significant performance improvement
(it saves 0 to 4 minutes for an update, but is actually used in very
rare cases such as hash algorithm change; it was never used in normal
operation), however it complicates code and also has fatal issue which,
in case of >25% but not all projects updates, results in deletion of all
not updated projects. Thankfully it seem to never have fired in production.