Add rank for url/project relations
For each link assigned to a project, a rank is calculated as a
quotient of a spread (number of repository families) for a particular
link to a maximum spread.

That is, if some specific link is provided by all repository families
which define links for a particular project, it gets 1.0. If only one of
10 families provide it, it gets 0.1.

The rank would be used related projects calculation, to sink projects
discovered through rare links as less relevant.

Needed for repology/repology-webapp#74