Implement "incomplete" repository flag
This flag is to be used for repositories for which repology does
not see all the packages they actually contain. Examples:

- Stackage, which generates incomplete dumps: https://github.com/commercialhaskell/stackage/issues/5352
- OpenWRT master, which seem to list binary packages which frequently
  fail to build and thus disappear from the dump
- PyPI, for which we have to setup a custom service which generates a
  dump based on recently changed packages only

Normally, such repositories result in polluting feeds with a lot of
package added/removed events, but this flag disables such event
generation for affected repositories.

However, this flag does not prevent such repositories from generating
latest versions which may still blink and pollute event feeds, so in
some cases it may be worth using this along with ignoring all versions
from these repositories.