Switch centos to generated packagelinks
This needs new keyword in links formatter which expands to either
-extra or an empty string based on subrepo. The keyword is then
used to generate correct branch (e.g. c7 or c7-extra). Empty keyword
handling was also changed in LinkFormatter - we allow it for optional
keywords but handle it differently than Null values, not skipping
generated links in this case.