Allow selective extraction of Vcs-* links from debian package lists
In some repos a subset of these links may be valid, so allow to specify
regexp for allowed urls. Use this feature for kali linux, for which we
want to only extract PACKAGE_SOURCES links for kali-specific packages
and not ones which come from debian. Generating such links produced
97.5% of broken links since most packages are from Debian and don't have
repository at https://gitlab.com/kalilinux/packages, this is now fixed
too.