Optimize ruleset lookup
Currently, we were processing 'ruleset' and 'noruleset' rule
conditions in regular matchers, which could be quite slow because
it involved set intersection.

However, we always process packages for a single repository in a
batch, so we always know a set of rulesets beforehand. Given that,
we could prepare a subset of a ruleset for parsing individual
repository, with all 'ruleset'/'noruleset' conditions resolved.

This should be much faster, as not only it avoids slow set
intersections, but it complerelt skips a subset of rules from
processing.

This yields ~35% parsing performance gain and partially fixes
performance regression introduced in b1b9359 (after which we were
losing rule match statistics after parsing each repository, and
in addition rule match frequencies were calculated incorrectly
which led to suboptimal ruleset lookup strategies).