Add global version update log
Log all newest/devel versions changes, not tied to any maintainer or
repository.

This could be usable as a news feed, and for that purpose also log
spread to allow filtering out not widely spread projects. If that's
demed useful, partial index may be added for fast lookup, e.g.

  CREATE INDEX ON global_version_events(ts DESC) WHERE (spread > 3)

Also log repositories which triggered an update, this may be useful
for statistics, e.g. to find which repositories trigger updates the
most.