nix: Don't exclude team members
This partially reverts 3fb04fa66f02377c26dab6fd747168a38d60d361.

As noted in a comment on #1497 [1] the addition of `meta.teams` to
`meta.maintainers` was just a return to the status quo before the
introduction of separate field. In the long run it makes sense to handle
teams as single entities but while it isn't implemented let's leave them
as is, since some people rely on this [2].

FWIW the largest team in nixpkgs counts 10 members so most of the team
owned packages will likely still be under the limit.

[1]: https://github.com/repology/repology-updater/issues/1497#issuecomment-2899308391
[2]: https://github.com/NixOS/nixpkgs/pull/402991#issuecomment-2853221104

Signed-off-by: Marcin Serwin <marcin@serwin.dev>