Use link failure streak when generating problems
When generating problems for failed homepage and download links, in
addition to a month passed since link birth or last known success,
require at least 3 consequentional failures.

This should reduce number of false positives, as due to allowable
recheck rates it was possible to have just a single failed check
in a month, which could be sporadic, which, in turn, would lead to
a false positive problem. Requirement for 3 failures alleviates this
problem, and if --fast-failure-recheck is enabled in link checker,
problem discovery time will not suffer either.